pyproject.toml
[tool.poetry]
name = "articles"
version = "0.1.0"
description = "articles"
authors = ["AnnaKPolyakova"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.11"
django-rest-framework = "^0.1.0"
drf-spectacular = "^0.21.1"
pytest = "^6.2.5"
pytest-django = "^4.5.2"
isort = "^5.10.1"
gunicorn = "^20.1.0"
django-environ = "0.7.0"
pytest-lazy-fixture = "^0.6.3"
pytest-cov = "^3.0.0"
django-filter = "^21.1"
psycopg2-binary = "^2.9.3"
black = "^22.3.0"
flake8 = "^4.0.1"
